requirements-dev.txt
# Code Formatter
black==24.4.2

# Code Linter
flake8==7.0.0

# Testing Framework
pytest==8.2.2

# for env
python-dotenv==1.1.1

# for Database
sqlalchemy==2.0.41
mysql-connector-python==9.0.0